add: advanced debouncing
This should help greatly with debouncing via an advanced debouncing
algorithm

+++ keyboard-specific-builds/diverge-1/diverge-left/animus/EEPROM.ino
@@ -6,7 +6,7 @@ void SetEEPROM(int x, int y, int z, char inputChar, byte inputType)
   EEPROM.update(addr, inputType);
 }
 
-int GetLayEEPROM()
+byte GetLayEEPROM()
 {
   return EEPROM.read(1023);
 }
